+++ CHANGELOG.md
@@ -30,6 +30,7 @@
   method.
 - animate() returned promise is not an animation controller object from now on,
   but has a member promise called 'activated', which resolves to the controller.
+- New style parameter 'numberScale' introduced for setting big number formats.
 
 ## [0.6.1] - 2022-11-22